LPD-64513 Move duplicated logic to FragmentEntryMenuDisplayConfiguration.getSiteNavigationMenuId
# breaking

## What modules/apps/fragment/fragment-api/src/main/java/com/liferay/fragment/util/configuration/FragmentEntryMenuDisplayConfiguration.java

modules/apps/fragment/fragment-api/src/main/java/com/liferay/fragment/util/configuration/FragmentEntryMenuDisplayConfiguration.java has had the getSiteNavigationMenuId method updated to include the parameter groupId. All usages have been updated as well.

## Why

This change was made to make it possible to get the siteNavigationMenuId from the siteNavigationMenuExternalReferenceCode in cases where the siteNavigationMenuId is not known such as when the configuration is imported via a site initializer. The groupId is needed to fetch the navigationMenu by the externalReferenceCode.

----